+++ internal/logger/logger.go
@@ -43,11 +43,12 @@ type Msg struct {
 }
 
 type MsgLocation struct {
-	File     string
-	Line     int // 1-based
-	Column   int // 0-based, in bytes
-	Length   int // in bytes
-	LineText string
+	File      string
+	Namespace string
+	Line      int // 1-based
+	Column    int // 0-based, in bytes
+	Length    int // in bytes
+	LineText  string
 }
 
 type Loc struct {
@@ -523,7 +524,7 @@ loop:
 	return
 }
 
-func locationOrNil(source *Source, r Range) *MsgLocation {
+func LocationOrNil(source *Source, r Range) *MsgLocation {
 	if source == nil {
 		return nil
 	}
@@ -708,7 +709,7 @@ func (log Log) AddError(source *Source, loc Loc, text string) {
 	log.AddMsg(Msg{
 		Kind:     Error,
 		Text:     text,
-		Location: locationOrNil(source, Range{Loc: loc}),
+		Location: LocationOrNil(source, Range{Loc: loc}),
 	})
 }
 
@@ -716,7 +717,7 @@ func (log Log) AddWarning(source *Source, loc Loc, text string) {
 	log.AddMsg(Msg{
 		Kind:     Warning,
 		Text:     text,
-		Location: locationOrNil(source, Range{Loc: loc}),
+		Location: LocationOrNil(source, Range{Loc: loc}),
 	})
 }
 
@@ -724,7 +725,7 @@ func (log Log) AddRangeError(source *Source, r Range, text string) {
 	log.AddMsg(Msg{
 		Kind:     Error,
 		Text:     text,
-		Location: locationOrNil(source, r),
+		Location: LocationOrNil(source, r),
 	})
 }
 
@@ -732,6 +733,6 @@ func (log Log) AddRangeWarning(source *Source, r Range, text string) {
 	log.AddMsg(Msg{
 		Kind:     Warning,
 		Text:     text,
-		Location: locationOrNil(source, r),
+		Location: LocationOrNil(source, r),
 	})
 }