+++ lib/browser.ts
@@ -62,6 +62,7 @@ export const startService: typeof types.startService = options => {
       writeToStdin(bytes) {
         worker.postMessage(bytes)
       },
+      isSync: false,
     })
 
     return {