+++ pkg/cli/cli_impl.go
@@ -8,6 +8,7 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/evanw/esbuild/internal/helpers"
 	"github.com/evanw/esbuild/internal/logger"
 	"github.com/evanw/esbuild/pkg/api"
 )
@@ -188,7 +189,7 @@ func parseOptionsImpl(osArgs []string, buildOpts *api.BuildOptions, transformOpt
 				return fmt.Errorf("Missing \"=\": %q", value)
 			}
 			ext, text := value[:equals], value[equals+1:]
-			loader, err := parseLoader(text)
+			loader, err := helpers.ParseLoader(text)
 			if err != nil {
 				return err
 			}
@@ -196,7 +197,7 @@ func parseOptionsImpl(osArgs []string, buildOpts *api.BuildOptions, transformOpt
 
 		case strings.HasPrefix(arg, "--loader="):
 			value := arg[len("--loader="):]
-			loader, err := parseLoader(value)
+			loader, err := helpers.ParseLoader(value)
 			if err != nil {
 				return err
 			}
@@ -417,36 +418,6 @@ outer:
 			"Invalid target: %q (valid: esN, "+strings.Join(engines, ", ")+")", value)
 	}
 	return
-}
-
-func parseLoader(text string) (api.Loader, error) {
-	switch text {
-	case "js":
-		return api.LoaderJS, nil
-	case "jsx":
-		return api.LoaderJSX, nil
-	case "ts":
-		return api.LoaderTS, nil
-	case "tsx":
-		return api.LoaderTSX, nil
-	case "css":
-		return api.LoaderCSS, nil
-	case "json":
-		return api.LoaderJSON, nil
-	case "text":
-		return api.LoaderText, nil
-	case "base64":
-		return api.LoaderBase64, nil
-	case "dataurl":
-		return api.LoaderDataURL, nil
-	case "file":
-		return api.LoaderFile, nil
-	case "binary":
-		return api.LoaderBinary, nil
-	default:
-		return 0, fmt.Errorf("Invalid loader: %q (valid: "+
-			"js, jsx, ts, tsx, css, json, text, base64, dataurl, file, binary)", text)
-	}
 }
 
 // This returns either BuildOptions, TransformOptions, or an error